Moved the chart to the modal
